src_crawler/pyproject.toml
[project]
name = "prices-timeline"
version = "0.1.0"
description = "Web crawler for extracting promotional product deals from kupi.cz"
requires-python = ">=3.10"
dependencies = [
    "crawl4ai>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "mypy>=1.8.0",
]

[tool.uv]
# Use uv-managed Python
python-preference = "managed"
